refactor: 선택된 폴더 렌더링하는 코드 컴포넌트로 따로 분리
