+++ src/components/FloatingAddBtn.css
@@ -1,3 +1,7 @@
+.floating-add-btn {
+  display: none;
+}
+
 @media only screen and (max-width: 767px) {
   .floating-add-btn {
     display: flex;
