+++ src/components/FolderCard.css
@@ -4,8 +4,6 @@
   flex-wrap: wrap;
   justify-content: space-between;
   padding: 0;
-  background-color: #f5f5f5;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
 }
 
 .folder-card .link-card {
