+++ src/components/FolderSortingTab.js
@@ -1,16 +1,9 @@
 import { useEffect, useState } from 'react';
 import { getSortingTab, getSearchedFolders } from '../api';
-import { getTimeAgo } from '../getTimeAgo';
 import FolderCard from './FolderCard';
 import FolderHeader from './FolderHeader';
-import FloatingAddBtn from './FloatingAddBtn';
+import SelectedFolderRender from './SelectedFolderRender';
 import './FolderSortingTab.css';
-import './FloatingAddBtn.css';
-
-function formatDate(value) {
-  const date = new Date(value);
-  return `${date.getFullYear()}. ${date.getMonth() + 1}. ${date.getDate()}`;
-}
 
 function FolderSortingTab() {
   const [tab, setTab] = useState();
@@ -55,41 +48,6 @@ function FolderSortingTab() {
     );
   };
 
-  const selectedFolderRender = selectedFolder => {
-    return (
-      <ul className="folder-card">
-        {selectedFolder?.map(link => (
-          <li className="link-card" key={link.id}>
-            <a href={link.url}>
-              <img
-                className="card-img"
-                src={link.image_source || process.env.PUBLIC_URL + '/images/no-image.png'}
-                alt="카드 이미지"
-              />
-              <div className="info-area">
-                <div className="time-ago-area">
-                  <span className="time-ago">{getTimeAgo(link.created_at)}</span>
-                  <button
-                    onClick={e => {
-                      e.preventDefault();
-                    }}>
-                    <img className="kebab-btn" src={process.env.PUBLIC_URL + '/images/kebab.png'} alt="케밥 메뉴" />
-                  </button>
-                </div>
-                <p className="title">{link.title}</p>
-                <span className="date">{formatDate(link.created_at)}</span>
-              </div>
-            </a>
-            <button>
-              <img className="star-btn" src={process.env.PUBLIC_URL + '/images/star.png'} alt="즐겨 찾기" />
-            </button>
-          </li>
-        ))}
-        <FloatingAddBtn />
-      </ul>
-    );
-  };
-
   return (
     <>
       <div className="sorting-area">
@@ -118,7 +76,9 @@ function FolderSortingTab() {
 
       {selectedFolder === undefined && <FolderCard />}
       {selectedFolder !== undefined && selectedFolder.length === 0 && nothingDataRender()}
-      {selectedFolder !== undefined && selectedFolder.length > 0 && selectedFolderRender(selectedFolder)}
+      {selectedFolder !== undefined && selectedFolder.length > 0 && (
+        <SelectedFolderRender selectedFolder={selectedFolder} />
+      )}
     </>
   );
 }
